all genre endpoints working

+++ app/routes/GenreRoutes.php
@@ -13,7 +13,6 @@ $app->get('/genres', function (Request $request, Response $response) use ($genre
 
     $response = $response->withHeader('Content-Type', 'application/json');
     $response->getBody()->write(json_encode($genres));
-
     return $response->withStatus(200);
 });
 
@@ -21,37 +20,46 @@ $app->post('/genres', function (Request $request, Response $response) use ($genr
     $data = json_decode($request->getBody()->getContents(), true);
 
     if (empty($data['nombre'])) {
-        return $response->withStatus(400)->getBody()->write('El campo "nombre" no puede estar vacio');
+        $response->getBody()->write('Ingrese un nombre valido');
+        return $response->withStatus(400);
     }
 
     $genre = $genreController->createGenre($data['nombre']);
 
     $response = $response->withHeader('Content-Type', 'application/json');
     $response->getBody()->write(json_encode($genre));
-
     return $response->withStatus(201);
 });
 
 $app->put('/genres/{id}', function (Request $request, Response $response, array $args) use ($genreController) {
     $id = $args['id'];
-    $data = $request->getParsedBody();
+    $data = json_decode($request->getBody()->getContents(), true);
+
+    if(!$genreController->existsById($id)){
+        $response->getBody()->write("No existe elemento con el id numero ". $id);
+        return $response->withStatus(404);
+    }
 
     if (empty($data['nombre'])) {
-        throw new Exception('El campo "nombre" es requerido');
+        $response->getBody()->write('Ingrese un nombre valido');
+        return $response->withStatus(400);
     }
 
     $genre = $genreController->updateGenre($id, $data['nombre']);
 
     $response = $response->withHeader('Content-Type', 'application/json');
     $response->getBody()->write(json_encode($genre));
-
     return $response->withStatus(200);
 });
 
 $app->delete('/genres/{id}', function (Request $request, Response $response, array $args) use ($genreController) {
     $id = $args['id'];
 
-    $genreController->deleteGenre($id);
+    if(!$genreController->existsById($id)){
+        $response->getBody()->write("No existe elemento con el id numero ". $id);
+        return $response->withStatus(404);
+    }
 
+    $genreController->deleteGenre($id);
     return $response->withStatus(204);
 });
